+++ derp/derpserver/testdata/example_ss.txt
